#Entity Framework except ADD

+++ CleanArch.Core/IRepositores/IRepository.cs
@@ -3,9 +3,11 @@
     public interface IRepository<T> where T : class
     {
         Task<IReadOnlyList<T>> GetAllAsync();
-        Task<T> GetByIdAsync(long id);
+       // Task<T> GetByIdAsync(long id);
+        Task<T> GetByIdAsync(int id);
         Task<string> AddAsync(T entity);
         Task<string> UpdateAsync(T entity);
-        Task<string> DeleteAsync(long id);
+        //Task<string> DeleteAsync(long id);
+        Task<string> DeleteAsync(int id);
     }
 }
